Attempt to solve the privacy problems with PTS subscribers data
- drop the pts_public table completely. PTS subscribers data isn't available
  from non-DDs anymore. Using the guestdd account is required to access it.
- change the CGI to use the guestdd account. The rest of the functionality of
  the CGI isn't changed.

Consequences:
The pts_public table (which contained the (package, md5(email)) list) doesn't
exist anymore. Only the pts table (which contains the non-hashed version
remains, but this table is only accessible using the guestdd account, whose
password is only available to DDs.
The CGI still exists, and still provide the same information. I still believe
that the information it provides is reasonable (see
http://lists.debian.org/debian-project/2010/03/msg00011.html).
It is harder to use that data from !samosa.debian.org, but we can probably
live with it.


git-svn-id: svn+ssh://svn.debian.org/svn/collab-qa/udd@1709 3b15d4d3-bb24-0410-9696-dc0fab150647