Update regex to parse values from newer Lintian output
Lintian's output will soon output tags in the format

  C: pkg type (version) [arch]: tag extra

as opposed to

  C: pkg type: tag extra

that is currently used. Update the parser to support both formats. Thanks to
Niels Thykier for supplying the original regular expression.

TODO: Add new pkg_version pkg_arch columns to table to make use of this
additional data (so we know which package was tested).

Bug #647917.



git-svn-id: svn+ssh://svn.debian.org/svn/collab-qa/udd@2094 3b15d4d3-bb24-0410-9696-dc0fab150647