reset autoremoval counter when bug updated
This means the bug isn't removed from the autoremoval list anymore on every
update. The old way caused confusion, because any update of the bug resulted
in a 15 day delay before the bug was on the list again. This gave the
impression that the bug was no longer an issue. Now we just reset thet 15 day
counter when the bug is updated.